the site is deployed in firbase and the server vercel deployed side's url is given in the client side code to connect the server side

+++ src/Pages/CartDetails.jsx
@@ -1,26 +1,45 @@
-import { useState } from "react";
-import { useLoaderData } from "react-router-dom";
+import swal from "sweetalert";
 
 
 // eslint-disable-next-line react/prop-types
-const CartDetails = ({ cart,cartInfo }) => {
-    const [newCart,setNewCart] = useState(cartInfo)
+const CartDetails = ({ cart,setNewCart,newCart }) => {
+    
     // eslint-disable-next-line react/prop-types
-    const { brandName, image, name, shortDescription,_id} = cart
-    const handleToDelete = (_id)=>{
-        fetch(`http://localhost:5000/carts/${_id}`,{
-            method: 'DELETE',
-            headers: {
-                'content-type': 'application/json'
-            },
-            body: JSON.stringify(newCart)
+    const { brandName, image, name, shortDescription, _id } = cart
+    const handleToDelete = (_id) => {
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you will not be able to recover this file!",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
         })
-        .then(res => res.json())
-        .then(data => {
-            console.log(data);
-            const remaining = newCart.filter(carts => carts._id !== _id)
-            setNewCart(remaining);
-        })
+            .then((willDelete) => {
+                if (willDelete) {
+                    fetch(`https://brand-shop-server-flame-alpha.vercel.app/carts/${_id}`, {
+                        method: 'DELETE',
+                        headers: {
+                            'content-type': 'application/json'
+                        },
+                        body: JSON.stringify(cart)
+                    })
+                        .then(res => res.json())
+                        .then(data => {
+                            console.log(data);
+                            if (data.deletedCount > 0) {
+                                swal("Poof! Your file has been deleted!", {
+                                    icon: "success",
+                                });
+                            }
+                            // eslint-disable-next-line react/prop-types
+                            const remaining = newCart.filter(carts => carts._id !== _id)
+                            setNewCart(remaining);
+                        })
+                } else {
+                    swal("Your file is safe!");
+                }
+            });
+
     }
     return (
         <div>
@@ -31,7 +50,7 @@ const CartDetails = ({ cart,cartInfo }) => {
                     <h2 className="card-title mt-5 text-orange-400 text-2xl font-medium">{name}</h2>
                     <p className="mt-5">{shortDescription}</p>
                     <div className="card-actions justify-start my-5">
-                        <button onClick={()=>handleToDelete(_id)} className="btn bg-orange-400 text-white">delete</button>
+                        <button onClick={() => handleToDelete(_id)} className="btn bg-orange-400 text-white">delete</button>
                     </div>
                 </div>
             </div>
